# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TechShop E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: categoryId
          schema: { type: integer }
        - in: query
          name: minPrice
          schema: { type: number }
        - in: query
          name: maxPrice
          schema: { type: number }
        - in: query
          name: sortBy
          schema: { type: string }
        - in: query
          name: featured
          schema: { type: boolean }
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/{id}/stock:
    patch:
      operationId: updateStock
      tags: [products]
      summary: Update product stock
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}/variants:
    get:
      operationId: listProductVariants
      tags: [products]
      summary: List variants for a product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product variants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductVariant"
    post:
      operationId: createProductVariant
      tags: [products]
      summary: Create a variant for a product (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductVariantInput"
      responses:
        "201":
          description: Created variant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductVariant"
  /products/{id}/variants/{variantId}:
    patch:
      operationId: updateProductVariant
      tags: [products]
      summary: Update a product variant (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: variantId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductVariantUpdate"
      responses:
        "200":
          description: Updated variant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductVariant"
    delete:
      operationId: deleteProductVariant
      tags: [products]
      summary: Delete a product variant (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: variantId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Cart
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    Product:
      type: object
      required: [id, name, slug, price, stock, categoryId, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        originalPrice: { type: ["number", "null"] }
        discountPercent: { type: ["number", "null"] }
        stock: { type: integer }
        imageUrl: { type: ["string", "null"] }
        images: { type: ["string", "null"] }
        categoryId: { type: integer }
        categoryName: { type: ["string", "null"] }
        brand: { type: ["string", "null"] }
        sku: { type: ["string", "null"] }
        isFeatured: { type: boolean }
        isActive: { type: boolean }
        soldCount: { type: integer }
        rating: { type: ["number", "null"] }
        reviewCount: { type: integer }
        createdAt: { type: string }
    ProductInput:
      type: object
      required: [name, slug, price, stock, categoryId]
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        price: { type: number }
        originalPrice: { type: number }
        stock: { type: integer }
        imageUrl: { type: string }
        images: { type: string }
        categoryId: { type: integer }
        brand: { type: string }
        sku: { type: string }
        isFeatured: { type: boolean }
        isActive: { type: boolean }
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        price: { type: number }
        originalPrice: { type: number }
        stock: { type: integer }
        imageUrl: { type: string }
        images: { type: string }
        categoryId: { type: integer }
        brand: { type: string }
        sku: { type: string }
        isFeatured: { type: boolean }
        isActive: { type: boolean }
    StockUpdate:
      type: object
      required: [stock]
      properties:
        stock: { type: integer }
        operation: { type: string, enum: [set, add, subtract] }
    ProductVariant:
      type: object
      required: [id, productId, name, value, priceModifier, stock, isDefault, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        name: { type: string }
        value: { type: string }
        priceModifier: { type: number }
        stock: { type: integer }
        sku: { type: ["string", "null"] }
        isDefault: { type: boolean }
        createdAt: { type: string }
    ProductVariantInput:
      type: object
      required: [name, value]
      properties:
        name: { type: string }
        value: { type: string }
        priceModifier: { type: number }
        stock: { type: integer }
        sku: { type: string }
        isDefault: { type: boolean }
    ProductVariantUpdate:
      type: object
      properties:
        name: { type: string }
        value: { type: string }
        priceModifier: { type: number }
        stock: { type: integer }
        sku: { type: string }
        isDefault: { type: boolean }
    ProductListResponse:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...