# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TechShop E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  slug: string;
  description?: string;
  price: number;
  originalPrice?: number;
  stock: number;
  imageUrl?: string;
  images?: string;
  categoryId: number;
  brand?: string;
  sku?: string;
  isFeatured?: boolean;
  isActive?: boolean;
}